ISSUE #84 * Minor refactor.
