+++ minos/common/model/declarative/aggregate/__init__.py
@@ -5,9 +5,9 @@ This file is part of minos framework.
 
 Minos framework can not be copied and/or distributed without the express permission of Clariteia SL.
 """
-from .abc import (
-    DynamicModel,
-)
 from .diff import (
-    FieldsDiff,
+    AggregateDiff,
+)
+from .impl import (
+    Aggregate,
 )
